+++ src/apis/axiosInstance.ts
@@ -3,6 +3,7 @@ import { z } from 'zod';
 
 export const axiosInstance = axios.create({
   baseURL: import.meta.env.VITE_API_HOST,
+  withCredentials: true,
   headers: {
     Authorization: import.meta.env.VITE_API_TOKEN,
   },
