+++ src/components/side-bar/SideBar.tsx
@@ -22,24 +22,6 @@ type MenubarProps = {
   setShowSidebar: React.Dispatch<React.SetStateAction<boolean>>;
 };
 
-function decodeJwt(token: string) {
-  try {
-    const payload = token.split('.')[1];
-    const decoded = JSON.parse(atob(payload));
-    return decoded;
-  } catch (error) {
-    console.error('JWT 디코딩 실패:', error);
-    return null;
-  }
-}
-
-const token = localStorage.getItem('accessToken');
-
-if (token) {
-  const decoded = decodeJwt(token);
-  console.log('디코딩된 payload:', decoded);
-}
-
 const SideBar: React.FC<MenubarProps> = ({
   avatarUrl,
   nickname,
