+++ src/routes/Router.tsx
@@ -3,10 +3,10 @@ import Layout from '../layout/layout';
 import LoginPage from '@/pages/pages/auth/login';
 import SignupPage from '@/pages/pages/auth/signup';
 import { LandingPage } from '@/pages/pages/landing/ui/LandingPage';
-import PageLayout from '@/components/page-layout/PageLayout';
 import ReissuePage from '@/pages/pages/reissue/page';
 import { ProtectedRoute } from './guards/ProtectedRoute';
 import { RedirectIfAuthenticated } from './guards/RedirectIfAuthenticated';
+import PersonalPage from '@/pages/PersonalPage';
 
 const router = createBrowserRouter([
   {
@@ -15,7 +15,7 @@ const router = createBrowserRouter([
       // 인증이 필요한 라우트들
       {
         element: <ProtectedRoute />,
-        children: [{ path: '/', element: <PageLayout /> }],
+        children: [{ path: '/', element: <PersonalPage /> }],
       },
 
       // 인증된 사용자는 접근할 필요 없는 라우트들
